refactor: class name, field name and etc
- jwt 코드 병합 내용 정리
