--- node-connection-backend/src/main/java/node/connection/_core/config/EncryptionConfig.java
@@ -1,15 +1,0 @@
-package node.connection._core.config;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-
-@Configuration
-public class EncryptionConfig {
-
-    @Bean
-    public BCryptPasswordEncoder passwordEncoder() {
-        return new BCryptPasswordEncoder();
-    }
-
-}
