--- node-connection-backend/src/main/java/node/connection/entity/UserEntity.java
@@ -1,24 +1,0 @@
-package node.connection.entity;
-
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import lombok.Getter;
-import lombok.Setter;
-
-@Entity
-@Setter
-@Getter
-public class UserEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
-
-    private String username;
-    private String password;
-    private String phonenum;
-
-    private String role;
-}
