--- node-connection-backend/src/main/java/node/connection/service/JoinService.java
@@ -1,42 +1,0 @@
-package node.connection.service;
-
-
-import node.connection.dto.user.request.JoinDTO;
-import node.connection.entity.UserEntity;
-import node.connection.repository.UserRepository;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import org.springframework.stereotype.Service;
-
-@Service
-public class JoinService {
-
-    private final UserRepository userRepository;
-    private final BCryptPasswordEncoder bCryptPasswordEncoder;
-
-    public JoinService(UserRepository userRepository, BCryptPasswordEncoder bCryptPasswordEncoder) {
-
-        this.userRepository = userRepository;
-        this.bCryptPasswordEncoder = bCryptPasswordEncoder;
-    }
-
-    public void joinProcess(JoinDTO joinDTO) {
-
-        String username = joinDTO.getUsername();
-        String password = joinDTO.getPassword();
-        String phonenum = joinDTO.getPhonenum();
-        Boolean isExist = userRepository.existsByUsername(username);
-
-        if (isExist) {
-            return;
-        }
-
-        UserEntity data = new UserEntity();
-
-        data.setUsername(username);
-        data.setPassword(bCryptPasswordEncoder.encode(password));
-        data.setPhonenum(phonenum);
-        data.setRole("ROLE_ADMIN");
-
-        userRepository.save(data);
-    }
-}
